+++ script.audio.motherearth/service.py
@@ -4,7 +4,7 @@ import xbmc
 import xbmcaddon
 import xbmcgui
 from collections import namedtuple
-from motherearth import STREAM_INFO, NowPlaying
+from motherearth import STREAM_INFO, NowPlaying, FANART_URL
 
 
 RESTART_INTERVAL = 1.0
@@ -12,7 +12,6 @@ RESTART_TIMEOUT = 1.0
 
 
 Song = namedtuple('Song', 'data cover')
-
 
 class Player(xbmc.Player):
     """Adds xbmc.Player callbacks and integrates with the API."""
@@ -87,7 +86,8 @@ class Player(xbmc.Player):
             item = xbmcgui.ListItem()
             item.setPath(self.getPlayingFile())
             item.setArt({'thumb': song.cover})
-            item.setArt({'fanart': song.cover})
+ #           item.setArt({'fanart': song.cover})
+            item.setArt({'fanart' : FANART_URL.format((song.data['artist']))})
             item.setInfo('music', info)
             self.updateInfoTag(item)